Replaced old consensus tool
