+++ simulator_tests/__init__.py
@@ -13,7 +13,6 @@ from .test_codereview_validation import CodeReviewValidationTest
 from .test_consensus_conversation import TestConsensusConversation
 from .test_consensus_stance import TestConsensusStance
 from .test_consensus_three_models import TestConsensusThreeModels
-from .test_consensusworkflow_validation import ConsensusWorkflowValidationTest
 from .test_content_validation import ContentValidationTest
 from .test_conversation_chain_validation import ConversationChainValidationTest
 from .test_cross_tool_comprehensive import CrossToolComprehensiveTest
@@ -73,7 +72,6 @@ TEST_REGISTRY = {
     "consensus_conversation": TestConsensusConversation,
     "consensus_stance": TestConsensusStance,
     "consensus_three_models": TestConsensusThreeModels,
-    "consensusworkflow_validation": ConsensusWorkflowValidationTest,
     "analyze_validation": AnalyzeValidationTest,
     # "o3_pro_expensive": O3ProExpensiveTest,  # COMMENTED OUT - too expensive to run by default
 }
@@ -110,7 +108,6 @@ __all__ = [
     "TestConsensusConversation",
     "TestConsensusStance",
     "TestConsensusThreeModels",
-    "ConsensusWorkflowValidationTest",
     "AnalyzeValidationTest",
     "TEST_REGISTRY",
 ]
