--- simulator_tests/test_consensusworkflow_validation.py
@@ -1,606 +1,0 @@
-#!/usr/bin/env python3
-"""
-ConsensusWorkflow Tool Validation Test
-
-Tests the consensus workflow tool's capabilities for multi-model consensus gathering.
-This validates the step-by-step workflow where Claude provides initial analysis,
-then consults each specified model, and finally synthesizes all perspectives.
-"""
-
-import json
-from typing import Optional
-
-from .conversation_base_test import ConversationBaseTest
-
-
-class ConsensusWorkflowValidationTest(ConversationBaseTest):
-    """Test consensus workflow tool with multi-model analysis"""
-
-    @property
-    def test_name(self) -> str:
-        return "consensusworkflow_validation"
-
-    @property
-    def test_description(self) -> str:
-        return "Consensus workflow tool validation with multi-model step-by-step analysis"
-
-    def run_test(self) -> bool:
-        """Test consensus workflow tool capabilities"""
-        # Set up the test environment
-        self.setUp()
-
-        try:
-            self.logger.info("Test: ConsensusWorkflow tool validation")
-
-            # Create test proposal files for consensus analysis
-            self._create_proposal_files()
-
-            # Test 1: Basic consensus workflow with 2 models
-            if not self._test_basic_consensus_workflow():
-                return False
-
-            # Test 2: Consensus with different stances (for/against)
-            if not self._test_consensus_with_stances():
-                return False
-
-            # Test 3: Consensus with file context
-            if not self._test_consensus_with_files():
-                return False
-
-            # Test 4: Multi-model consensus (3+ models)
-            if not self._test_multi_model_consensus():
-                return False
-
-            # Test 5: Duplicate model+stance validation
-            if not self._test_duplicate_model_stance_validation():
-                return False
-
-            self.logger.info("  ✅ All consensus workflow validation tests passed")
-            return True
-
-        except Exception as e:
-            self.logger.error(f"ConsensusWorkflow validation test failed: {e}")
-            return False
-
-    def _create_proposal_files(self):
-        """Create test proposal files for consensus analysis"""
-        # Create a feature proposal document
-        proposal_content = """# Feature Proposal: Real-time Collaborative Editing
-
-## Overview
-Implement real-time collaborative editing capabilities in our document editor, similar to Google Docs.
-
-## Technical Approach
-- Use WebSockets for real-time communication
-- Implement Operational Transformation (OT) algorithm for conflict resolution
-- Add presence awareness (see who's editing what)
-- Store document history for versioning
-
-## Benefits
-- Enhanced team collaboration
-- Reduced email/file sharing overhead
-- Better version control
-- Increased productivity
-
-## Challenges
-- Complex conflict resolution
-- Scalability concerns with many concurrent users
-- Increased server infrastructure costs
-- Potential security implications
-
-## Estimated Timeline
-- Phase 1 (Basic collaboration): 3 months
-- Phase 2 (Advanced features): 2 months
-- Total: 5 months with 2 developers
-"""
-
-        # Create architecture diagram reference
-        architecture_notes = """# Architecture Considerations
-
-## Current Architecture
-- Monolithic Django application
-- PostgreSQL database
-- Redis for caching
-- Nginx reverse proxy
-
-## Proposed Changes
-- Add WebSocket server (possibly separate service)
-- Implement message queue (RabbitMQ/Kafka)
-- Add real-time data synchronization layer
-- Consider microservices for collaboration features
-
-## Scaling Concerns
-- Need to handle 1000+ concurrent connections
-- Database write amplification with frequent saves
-- Network bandwidth for real-time updates
-- Client-side performance with large documents
-"""
-
-        self.proposal_file = self.create_additional_test_file("collaboration_proposal.md", proposal_content)
-        self.architecture_file = self.create_additional_test_file("architecture_notes.md", architecture_notes)
-        self.logger.info(f"  ✅ Created proposal files: {self.proposal_file}, {self.architecture_file}")
-
-    def _test_basic_consensus_workflow(self) -> bool:
-        """Test basic consensus workflow with 2 models"""
-        try:
-            self.logger.info("  1.1: Testing basic consensus workflow (2 models)")
-
-            # Step 1: Claude's initial analysis
-            self.logger.info("    1.1.1: Step 1 - Claude's initial analysis")
-            response1, continuation_id = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "I'll analyze this real-time collaboration proposal. The feature would add significant value for team productivity, but comes with substantial technical complexity. The OT algorithm implementation is non-trivial, and scaling WebSocket connections requires careful architecture. The 5-month timeline seems optimistic given the complexity.",
-                    "step_number": 1,
-                    "total_steps": 4,  # 1 (Claude) + 2 models + 1 (synthesis)
-                    "next_step_required": True,
-                    "findings": "Proposal offers strong collaboration benefits but involves complex technical challenges including OT implementation, WebSocket scaling, and infrastructure costs. Timeline may be underestimated.",
-                    "confidence": "medium",
-                    "models": [{"model": "flash", "stance": "neutral"}, {"model": "o3-mini", "stance": "neutral"}],
-                    "relevant_files": [self.proposal_file],
-                },
-            )
-
-            if not response1 or not continuation_id:
-                self.logger.error("Failed to get initial analysis response")
-                return False
-
-            # Parse and validate response
-            response1_data = self._parse_response(response1)
-            if not response1_data:
-                return False
-
-            # Validate step 1 response
-            if response1_data.get("status") != "consulting_models":
-                self.logger.error(f"Expected status 'consulting_models', got '{response1_data.get('status')}'")
-                return False
-
-            if response1_data.get("step_number") != 1:
-                self.logger.error(f"Expected step_number 1, got {response1_data.get('step_number')}")
-                return False
-
-            self.logger.info(f"    ✅ Step 1 successful, continuation_id: {continuation_id}")
-
-            # Step 2: First model consultation (should be done automatically by tool)
-            self.logger.info("    1.1.2: Step 2 - First model consultation")
-            response2, _ = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Processing first model response",
-                    "step_number": 2,
-                    "total_steps": 4,
-                    "next_step_required": True,
-                    "findings": "Flash model provided neutral analysis highlighting both benefits and risks. Key points: strong collaboration value, but OT complexity and scaling challenges are significant concerns.",
-                    "confidence": "medium",
-                    "continuation_id": continuation_id,
-                    "current_model_index": 0,
-                },
-            )
-
-            if not response2:
-                self.logger.error("Failed to get step 2 response")
-                return False
-
-            response2_data = self._parse_response(response2)
-
-            # Check for model consultation result
-            if response2_data.get("status") != "model_consulted":
-                self.logger.error(f"Expected status 'model_consulted', got '{response2_data.get('status')}'")
-                return False
-
-            model_response = response2_data.get("model_response", {})
-            if not model_response or model_response.get("status") != "success":
-                self.logger.error("Model consultation should have succeeded")
-                return False
-
-            self.logger.info(f"    ✅ Step 2: Consulted {model_response.get('model')} successfully")
-
-            # Step 3: Second model consultation
-            self.logger.info("    1.1.3: Step 3 - Second model consultation")
-            response3, _ = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Processing second model response",
-                    "step_number": 3,
-                    "total_steps": 4,
-                    "next_step_required": True,
-                    "findings": "O3-mini model also provided balanced perspective. Emphasized importance of phased approach and suggested considering lighter alternatives before full OT implementation.",
-                    "confidence": "high",
-                    "continuation_id": continuation_id,
-                    "current_model_index": 1,
-                },
-            )
-
-            if not response3:
-                self.logger.error("Failed to get step 3 response")
-                return False
-
-            response3_data = self._parse_response(response3)
-
-            if response3_data.get("next_step_required") is True:
-                self.logger.error("After consulting all models, next_step_required should be False")
-                return False
-
-            self.logger.info("    ✅ Step 3: Consulted second model successfully")
-
-            # Step 4: Final synthesis
-            self.logger.info("    1.1.4: Step 4 - Final synthesis")
-            response4, _ = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Synthesizing all perspectives for final recommendation",
-                    "step_number": 4,
-                    "total_steps": 4,
-                    "next_step_required": False,  # Final step
-                    "findings": "Consensus achieved: Feature has strong value proposition but implementation complexity is high. Both models and my analysis agree on phased approach starting with simpler collaboration features before full OT implementation.",
-                    "confidence": "high",
-                    "continuation_id": continuation_id,
-                },
-            )
-
-            if not response4:
-                self.logger.error("Failed to get final synthesis response")
-                return False
-
-            response4_data = self._parse_response(response4)
-
-            # Validate final response
-            if response4_data.get("status") != "consensus_workflow_complete":
-                self.logger.error(
-                    f"Expected status 'consensus_workflow_complete', got '{response4_data.get('status')}'"
-                )
-                return False
-
-            if not response4_data.get("consensus_complete"):
-                self.logger.error("Expected consensus_complete=true for final step")
-                return False
-
-            # Check complete consensus data
-            complete_consensus = response4_data.get("complete_consensus", {})
-            models_consulted = complete_consensus.get("models_consulted", [])
-
-            if len(models_consulted) < 2:
-                self.logger.error(f"Expected at least 2 models consulted, got {len(models_consulted)}")
-                return False
-
-            self.logger.info(f"    ✅ Consensus complete with {len(models_consulted)} models")
-            self.logger.info("    ✅ Basic consensus workflow test passed")
-            return True
-
-        except Exception as e:
-            self.logger.error(f"Basic consensus workflow test failed: {e}")
-            return False
-
-    def _test_consensus_with_stances(self) -> bool:
-        """Test consensus with different stances (for/against)"""
-        try:
-            self.logger.info("  1.2: Testing consensus with stances (for/against)")
-
-            # Step 1: Claude's initial analysis
-            self.logger.info("    1.2.1: Step 1 - Claude's analysis with stance models")
-            response1, continuation_id = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Analyzing the collaboration proposal with awareness that we'll get contrasting perspectives. The feature has clear benefits for team productivity but also significant technical and cost challenges.",
-                    "step_number": 1,
-                    "total_steps": 4,
-                    "next_step_required": True,
-                    "findings": "Balanced initial assessment recognizing both strong value proposition and implementation complexity. Will be interesting to see supportive vs critical perspectives.",
-                    "confidence": "medium",
-                    "models": [{"model": "flash", "stance": "for"}, {"model": "o3-mini", "stance": "against"}],
-                    "relevant_files": [self.proposal_file, self.architecture_file],
-                },
-            )
-
-            if not response1 or not continuation_id:
-                self.logger.error("Failed to start stance-based consensus")
-                return False
-
-            # Continue through the workflow steps
-            # Step 2: Pro stance model
-            response2, _ = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Processing supportive perspective",
-                    "step_number": 2,
-                    "total_steps": 4,
-                    "next_step_required": True,
-                    "findings": "Flash (for stance) emphasized strong ROI, competitive advantage, and how modern frameworks make implementation easier than anticipated.",
-                    "confidence": "medium",
-                    "continuation_id": continuation_id,
-                    "current_model_index": 0,
-                },
-            )
-
-            if not response2:
-                self.logger.error("Failed to process pro stance")
-                return False
-
-            # Step 3: Against stance model
-            response3, _ = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Processing critical perspective",
-                    "step_number": 3,
-                    "total_steps": 4,
-                    "next_step_required": True,
-                    "findings": "O3-mini (against stance) highlighted serious scalability risks, underestimated complexity, and suggested existing solutions might be more cost-effective.",
-                    "confidence": "high",
-                    "continuation_id": continuation_id,
-                    "current_model_index": 1,
-                },
-            )
-
-            if not response3:
-                self.logger.error("Failed to process against stance")
-                return False
-
-            # Step 4: Final synthesis with contrasting views
-            response4, _ = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Synthesizing contrasting perspectives",
-                    "step_number": 4,
-                    "total_steps": 4,
-                    "next_step_required": False,
-                    "findings": "Clear divide: supportive view sees transformative potential while critical view warns of implementation pitfalls. Consensus suggests proof-of-concept with limited scope first.",
-                    "confidence": "high",
-                    "continuation_id": continuation_id,
-                },
-            )
-
-            if not response4:
-                self.logger.error("Failed to complete stance synthesis")
-                return False
-
-            response4_data = self._parse_response(response4)
-
-            # Verify stances were properly recorded
-            complete_consensus = response4_data.get("complete_consensus", {})
-            models_consulted = complete_consensus.get("models_consulted", [])
-
-            if "flash:for" not in models_consulted:
-                self.logger.error("Expected flash:for in models consulted")
-                return False
-
-            if "o3-mini:against" not in models_consulted:
-                self.logger.error("Expected o3-mini:against in models consulted")
-                return False
-
-            self.logger.info("    ✅ Consensus with stances test passed")
-            return True
-
-        except Exception as e:
-            self.logger.error(f"Consensus with stances test failed: {e}")
-            return False
-
-    def _test_consensus_with_files(self) -> bool:
-        """Test consensus workflow with file context"""
-        try:
-            self.logger.info("  1.3: Testing consensus with file context")
-
-            # Create additional technical spec file
-            tech_spec = """# Technical Specification
-
-## WebSocket Implementation
-- Use Socket.IO for WebSocket abstraction
-- Implement heartbeat mechanism
-- Handle reconnection logic
-- Message queuing during disconnection
-
-## Operational Transformation
-- Character-based OT for text editing
-- Support for cursor positions
-- Undo/redo stack per user
-- Conflict resolution algorithms
-
-## Performance Requirements
-- < 100ms latency for local edits
-- < 500ms for remote updates
-- Support 50 concurrent editors per document
-- Document size limit: 10MB
-"""
-
-            tech_file = self.create_additional_test_file("tech_spec.md", tech_spec)
-
-            # Step 1: With multiple files
-            response1, continuation_id = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Analyzing detailed technical specifications alongside the proposal. The technical depth shows careful planning but also reveals the significant complexity involved.",
-                    "step_number": 1,
-                    "total_steps": 3,  # Shorter workflow for this test
-                    "next_step_required": True,
-                    "findings": "Technical specifications are thorough but confirm high implementation complexity. Performance requirements are aggressive for the proposed architecture.",
-                    "confidence": "medium",
-                    "models": [{"model": "flash", "stance": "neutral"}],
-                    "relevant_files": [self.proposal_file, self.architecture_file, tech_file],
-                },
-            )
-
-            if not response1 or not continuation_id:
-                self.logger.error("Failed to start file context consensus")
-                return False
-
-            response1_data = self._parse_response(response1)
-
-            # Files should be referenced in step 1
-            if "file_context" in response1_data:
-                self.logger.info("    ✅ File context properly handled in workflow")
-
-            # Continue through remaining steps...
-            # (Abbreviated for brevity - would normally test all steps)
-
-            self.logger.info("    ✅ Consensus with files test passed")
-            return True
-
-        except Exception as e:
-            self.logger.error(f"Consensus with files test failed: {e}")
-            return False
-
-    def _test_multi_model_consensus(self) -> bool:
-        """Test consensus with 3+ models"""
-        try:
-            self.logger.info("  1.4: Testing multi-model consensus (3+ models)")
-
-            # Step 1: Claude's analysis with 3 models
-            response1, continuation_id = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Analyzing proposal for comprehensive multi-model consensus. This will provide diverse perspectives on feasibility.",
-                    "step_number": 1,
-                    "total_steps": 5,  # 1 + 3 models + 1 synthesis
-                    "next_step_required": True,
-                    "findings": "Initial assessment for multi-perspective analysis. Collaboration features have clear value but implementation complexity varies based on approach.",
-                    "confidence": "medium",
-                    "models": [
-                        {"model": "flash", "stance": "for"},
-                        {"model": "o3-mini", "stance": "neutral"},
-                        {
-                            "model": "flash",
-                            "stance": "against",
-                            "stance_prompt": "Focus on technical debt and maintenance burden",
-                        },
-                    ],
-                    "relevant_files": [self.proposal_file],
-                },
-            )
-
-            if not response1 or not continuation_id:
-                self.logger.error("Failed to start multi-model consensus")
-                return False
-
-            # Process through all model consultations
-            model_count = 3
-            for i in range(model_count):
-                step_num = i + 2
-                self.logger.info(f"    1.4.{step_num}: Step {step_num} - Model {i+1} consultation")
-
-                response, _ = self.call_mcp_tool(
-                    "consensusworkflow",
-                    {
-                        "step": f"Processing model {i+1} response",
-                        "step_number": step_num,
-                        "total_steps": 5,
-                        "next_step_required": step_num < 4,  # True until last model
-                        "findings": f"Model {i+1} provided perspective on implementation approach",
-                        "confidence": "medium" if i < 2 else "high",
-                        "continuation_id": continuation_id,
-                        "current_model_index": i,
-                    },
-                )
-
-                if not response:
-                    self.logger.error(f"Failed to process model {i+1}")
-                    return False
-
-            # Final synthesis with 3 models
-            response_final, _ = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Synthesizing perspectives from 3 different models",
-                    "step_number": 5,
-                    "total_steps": 5,
-                    "next_step_required": False,
-                    "findings": "Three models provided diverse perspectives: supportive view emphasized benefits, neutral highlighted tradeoffs, critical focused on maintenance burden. Consensus: start small with MVP.",
-                    "confidence": "high",
-                    "continuation_id": continuation_id,
-                },
-            )
-
-            if not response_final:
-                self.logger.error("Failed to complete multi-model synthesis")
-                return False
-
-            response_final_data = self._parse_response(response_final)
-
-            # Verify all models were consulted
-            complete_consensus = response_final_data.get("complete_consensus", {})
-            total_responses = complete_consensus.get("total_responses", 0)
-
-            if total_responses < 3:
-                self.logger.error(f"Expected at least 3 model responses, got {total_responses}")
-                return False
-
-            self.logger.info(f"    ✅ Multi-model consensus with {total_responses} models passed")
-            return True
-
-        except Exception as e:
-            self.logger.error(f"Multi-model consensus test failed: {e}")
-            return False
-
-    def _test_duplicate_model_stance_validation(self) -> bool:
-        """Test that duplicate model+stance combinations are properly rejected"""
-        try:
-            self.logger.info("  1.5: Testing duplicate model+stance validation")
-
-            # This should fail due to duplicate o3:for combination
-            response, _ = self.call_mcp_tool(
-                "consensusworkflow",
-                {
-                    "step": "Testing duplicate validation",
-                    "step_number": 1,
-                    "total_steps": 1,
-                    "next_step_required": True,
-                    "findings": "Testing validation",
-                    "models": [
-                        {"model": "o3", "stance": "for"},
-                        {"model": "flash", "stance": "neutral"},
-                        {"model": "o3", "stance": "for"},  # Duplicate!
-                    ],
-                    "relevant_files": [self.proposal_file],
-                },
-            )
-
-            if response:
-                response_data = self._parse_response(response)
-                # Check if we got an error about duplicate
-                if response_data.get("error"):
-                    error_msg = str(response_data.get("error", ""))
-                    if "Duplicate model + stance combination" in error_msg:
-                        self.logger.info("    ✅ Duplicate model+stance correctly rejected")
-                        return True
-                    else:
-                        self.logger.error(f"Got unexpected error: {error_msg}")
-                        return False
-                else:
-                    self.logger.error("Expected duplicate validation error but request succeeded")
-                    return False
-            else:
-                # No response means the request was rejected
-                self.logger.info("    ✅ Duplicate model+stance validation working correctly")
-                return True
-
-        except Exception as e:
-            self.logger.error(f"Duplicate validation test failed: {e}")
-            return False
-
-    def call_mcp_tool(self, tool_name: str, params: dict) -> tuple[Optional[str], Optional[str]]:
-        """Call an MCP tool and extract continuation_id"""
-        response_text, _ = self.call_mcp_tool_direct(tool_name, params)
-
-        if not response_text:
-            return None, None
-
-        # Extract continuation_id from response
-        continuation_id = self._extract_continuation_id(response_text)
-
-        return response_text, continuation_id
-
-    def _extract_continuation_id(self, response_text: str) -> Optional[str]:
-        """Extract continuation_id from response"""
-        try:
-            response_data = json.loads(response_text)
-            return response_data.get("continuation_id")
-        except json.JSONDecodeError:
-            self.logger.debug("Failed to parse response for continuation_id")
-            return None
-
-    def _parse_response(self, response_text: str) -> dict:
-        """Parse consensus workflow response"""
-        try:
-            return json.loads(response_text)
-        except json.JSONDecodeError as e:
-            self.logger.error(f"Failed to parse response as JSON: {e}")
-            self.logger.error(f"Response text: {response_text[:500]}...")
-            return {}
