+++ tools/__init__.py
@@ -6,7 +6,6 @@ from .analyze import AnalyzeTool
 from .chat import ChatTool
 from .codereview import CodeReviewTool
 from .consensus import ConsensusTool
-from .consensusworkflow import ConsensusWorkflowTool
 from .debug import DebugIssueTool
 from .listmodels import ListModelsTool
 from .planner import PlannerTool
@@ -23,7 +22,6 @@ __all__ = [
     "AnalyzeTool",
     "ChatTool",
     "ConsensusTool",
-    "ConsensusWorkflowTool",
     "ListModelsTool",
     "PlannerTool",
     "PrecommitTool",
