+++ tools/consensus.py
@@ -1,358 +1,540 @@
 """
-Consensus tool for multi-model perspective gathering and validation
+Consensus tool - Step-by-step multi-model consensus with expert analysis
+
+This tool provides a structured workflow for gathering consensus from multiple models.
+It guides Claude through systematic steps where Claude first provides its own analysis,
+then consults each requested model one by one, and finally synthesizes all perspectives.
+
+Key features:
+- Step-by-step consensus workflow with progress tracking
+- Claude's initial neutral analysis followed by model-specific consultations
+- Context-aware file embedding
+- Support for stance-based analysis (for/against/neutral)
+- Final synthesis combining all perspectives
 """
+
+from __future__ import annotations
 
 import json
 import logging
-from typing import TYPE_CHECKING, Any, Optional
+from typing import TYPE_CHECKING, Any
 
-from mcp.types import TextContent
-from pydantic import BaseModel, Field, field_validator
+from pydantic import Field, model_validator
 
 if TYPE_CHECKING:
     from tools.models import ToolModelCategory
 
-from config import DEFAULT_CONSENSUS_MAX_INSTANCES_PER_COMBINATION
-from systemprompts import CONSENSUS_PROMPT
+from mcp.types import TextContent
 
-from .base import BaseTool, ToolRequest
+from config import TEMPERATURE_ANALYTICAL
+from systemprompts import CONSENSUS_PROMPT
+from tools.shared.base_models import WorkflowRequest
+
+from .workflow.base import WorkflowTool
 
 logger = logging.getLogger(__name__)
 
-# Field descriptions to avoid duplication between Pydantic and JSON schema
-CONSENSUS_FIELD_DESCRIPTIONS = {
-    "prompt": (
-        "Description of what to get consensus on, testing objectives, and specific scope/focus areas. "
-        "Be as detailed as possible about the proposal, plan, or idea you want multiple perspectives on."
+# Tool-specific field descriptions for consensus workflow
+CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS = {
+    "step": (
+        "Describe your current consensus analysis step. In step 1, provide your own neutral, balanced analysis "
+        "of the proposal/idea/plan after thinking carefully about all aspects. Consider technical feasibility, "
+        "user value, implementation complexity, and alternatives. In subsequent steps (2+), you will receive "
+        "individual model responses to synthesize. CRITICAL: Be thorough and balanced in your initial assessment, "
+        "considering both benefits and risks, opportunities and challenges."
+    ),
+    "step_number": (
+        "The index of the current step in the consensus workflow, beginning at 1. Step 1 is your analysis, "
+        "steps 2+ are for processing individual model responses."
+    ),
+    "total_steps": (
+        "Total number of steps needed. This equals 1 (your analysis) + number of models to consult + "
+        "1 (final synthesis)."
+    ),
+    "next_step_required": ("Set to true if more models need to be consulted. False when ready for final synthesis."),
+    "findings": (
+        "In step 1, provide your comprehensive analysis of the proposal. In steps 2+, summarize the key points "
+        "from the model response received, noting agreements and disagreements with previous analyses."
+    ),
+    "relevant_files": (
+        "Files that are relevant to the consensus analysis. Include files that help understand the proposal, "
+        "provide context, or contain implementation details."
     ),
     "models": (
-        "List of model configurations for consensus analysis. Each model can have a specific stance and custom instructions. "
-        "Example: [{'model': 'o3', 'stance': 'for', 'stance_prompt': 'Focus on benefits and opportunities...'}, "
-        "{'model': 'flash', 'stance': 'against', 'stance_prompt': 'Identify risks and challenges...'}]. "
-        "Maximum 2 instances per model+stance combination."
+        "List of model configurations to consult. Each can have a model name, stance (for/against/neutral), "
+        "and optional custom stance prompt. The same model can be used multiple times with different stances, "
+        "but each model + stance combination must be unique. "
+        "Example: [{'model': 'o3', 'stance': 'for'}, {'model': 'o3', 'stance': 'against'}, "
+        "{'model': 'flash', 'stance': 'neutral'}]"
     ),
-    "files": "Optional files or directories for additional context (must be FULL absolute paths - DO NOT SHORTEN)",
+    "current_model_index": (
+        "Internal tracking of which model is being consulted (0-based index). Used to determine which model "
+        "to call next."
+    ),
+    "model_responses": ("Accumulated responses from models consulted so far. Internal field for tracking progress."),
     "images": (
-        "Optional images showing expected UI changes, design requirements, "
-        "or visual references for the consensus analysis"
+        "Optional list of image paths or base64 data URLs for visual context. Useful for UI/UX discussions, "
+        "architecture diagrams, mockups, or any visual references that help inform the consensus analysis."
     ),
-    "focus_areas": "Specific aspects to focus on (e.g., 'performance', 'security', 'user experience')",
-    "model_config_model": "Model name to use (e.g., 'o3', 'flash', 'pro')",
-    "model_config_stance": (
-        "Stance for this model. Supportive: 'for', 'support', 'favor'. "
-        "Critical: 'against', 'oppose', 'critical'. Neutral: 'neutral'. "
-        "Defaults to 'neutral'."
-    ),
-    "model_config_stance_prompt": (
-        "Custom stance-specific instructions for this model. "
-        "If provided, this will be used instead of the default stance prompt. "
-        "Should be clear, specific instructions about how this model should approach the analysis."
-    ),
-    "model_config_stance_schema": "Stance for this model: supportive ('for', 'support', 'favor'), critical ('against', 'oppose', 'critical'), or 'neutral'",
 }
 
 
-class ModelConfig(BaseModel):
-    """Enhanced model configuration for consensus tool"""
-
-    model: str = Field(..., description=CONSENSUS_FIELD_DESCRIPTIONS["model_config_model"])
-    stance: Optional[str] = Field(
-        default="neutral",
-        description=CONSENSUS_FIELD_DESCRIPTIONS["model_config_stance"],
-    )
-    stance_prompt: Optional[str] = Field(
-        default=None,
-        description=CONSENSUS_FIELD_DESCRIPTIONS["model_config_stance_prompt"],
-    )
+class ModelConfig(dict):
+    """Model configuration for consensus workflow"""
 
 
-class ConsensusRequest(ToolRequest):
-    """Request model for consensus tool"""
+class ConsensusRequest(WorkflowRequest):
+    """Request model for consensus workflow steps"""
 
-    prompt: str = Field(..., description=CONSENSUS_FIELD_DESCRIPTIONS["prompt"])
-    models: list[ModelConfig] = Field(..., description=CONSENSUS_FIELD_DESCRIPTIONS["models"])
-    files: Optional[list[str]] = Field(
+    # Required fields for each step
+    step: str = Field(..., description=CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["step"])
+    step_number: int = Field(..., description=CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["step_number"])
+    total_steps: int = Field(..., description=CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["total_steps"])
+    next_step_required: bool = Field(..., description=CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["next_step_required"])
+
+    # Investigation tracking fields
+    findings: str = Field(..., description=CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["findings"])
+    confidence: str | None = Field("exploring", exclude=True)  # Not used in consensus workflow
+
+    # Consensus-specific fields (only needed in step 1)
+    models: list[dict] | None = Field(None, description=CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["models"])
+    relevant_files: list[str] | None = Field(
         default_factory=list,
-        description=CONSENSUS_FIELD_DESCRIPTIONS["files"],
+        description=CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["relevant_files"],
     )
-    images: Optional[list[str]] = Field(
+
+    # Internal tracking fields
+    current_model_index: int | None = Field(
+        0,
+        description=CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["current_model_index"],
+    )
+    model_responses: list[dict] | None = Field(
         default_factory=list,
-        description=CONSENSUS_FIELD_DESCRIPTIONS["images"],
-    )
-    focus_areas: Optional[list[str]] = Field(
-        default_factory=list,
-        description=CONSENSUS_FIELD_DESCRIPTIONS["focus_areas"],
+        description=CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["model_responses"],
     )
 
-    @field_validator("models")
-    @classmethod
-    def validate_models_not_empty(cls, v):
-        if not v:
-            raise ValueError("At least one model must be specified")
-        return v
+    # Override inherited fields to exclude them from schema
+    temperature: float | None = Field(default=None, exclude=True)
+    thinking_mode: str | None = Field(default=None, exclude=True)
+    use_websearch: bool | None = Field(default=None, exclude=True)
+
+    # Not used in consensus workflow
+    files_checked: list[str] | None = Field(default_factory=list, exclude=True)
+    relevant_context: list[str] | None = Field(default_factory=list, exclude=True)
+    issues_found: list[dict] | None = Field(default_factory=list, exclude=True)
+    hypothesis: str | None = Field(None, exclude=True)
+    backtrack_from_step: int | None = Field(None, exclude=True)
+    images: list[str] | None = Field(default_factory=list)  # Enable images for consensus workflow
+
+    @model_validator(mode="after")
+    def validate_step_one_requirements(self):
+        """Ensure step 1 has required models field and unique model+stance combinations."""
+        if self.step_number == 1:
+            if not self.models:
+                raise ValueError("Step 1 requires 'models' field to specify which models to consult")
+
+            # Check for unique model + stance combinations
+            seen_combinations = set()
+            for model_config in self.models:
+                model_name = model_config.get("model", "")
+                stance = model_config.get("stance", "neutral")
+                combination = f"{model_name}:{stance}"
+
+                if combination in seen_combinations:
+                    raise ValueError(
+                        f"Duplicate model + stance combination found: {model_name} with stance '{stance}'. "
+                        f"Each model + stance combination must be unique."
+                    )
+                seen_combinations.add(combination)
+
+        return self
 
 
-class ConsensusTool(BaseTool):
-    """Multi-model consensus tool for gathering diverse perspectives on technical proposals"""
+class ConsensusTool(WorkflowTool):
+    """
+    Consensus workflow tool for step-by-step multi-model consensus gathering.
+
+    This tool implements a structured consensus workflow where Claude first provides
+    its own neutral analysis, then consults each specified model individually,
+    and finally synthesizes all perspectives into a unified recommendation.
+    """
 
     def __init__(self):
         super().__init__()
+        self.initial_prompt: str | None = None
+        self.models_to_consult: list[dict] = []
+        self.accumulated_responses: list[dict] = []
+        self._current_arguments: dict[str, Any] = {}
 
     def get_name(self) -> str:
         return "consensus"
 
     def get_description(self) -> str:
         return (
-            "MULTI-MODEL CONSENSUS - Gather diverse perspectives from multiple AI models on project proposals, "
-            "feature designs, architectural decisions, and strategic plans. NOT for code review - use the dedicated "
-            "codereview tool instead. Perfect for: project planning validation, feature feasibility assessment, "
-            "technology choice decisions, product roadmap evaluation, and getting multiple viewpoints on high-level "
-            "design choices. Supports advanced stance steering with custom instructions for each model. "
-            "You can specify different stances (for/against/neutral) and provide custom stance prompts to guide each "
-            "model's analysis. Example: [{'model': 'o3', 'stance': 'for', 'stance_prompt': 'Focus on implementation "
-            "benefits and user value'}, {'model': 'flash', 'stance': 'against', 'stance_prompt': 'Identify potential "
-            "risks and technical challenges'}]. Use neutral stances by default unless structured debate would add value."
+            "COMPREHENSIVE CONSENSUS WORKFLOW - Step-by-step multi-model consensus with structured analysis. "
+            "This tool guides you through a systematic process where you:\\n\\n"
+            "1. Start with step 1: provide your own neutral analysis of the proposal\\n"
+            "2. The tool will then consult each specified model one by one\\n"
+            "3. You'll receive each model's response in subsequent steps\\n"
+            "4. Track and synthesize perspectives as they accumulate\\n"
+            "5. Final step: present comprehensive consensus and recommendations\\n\\n"
+            "IMPORTANT: This workflow enforces sequential model consultation:\\n"
+            "- Step 1 is always your independent analysis\\n"
+            "- Each subsequent step processes one model response\\n"
+            "- Total steps = 1 (your analysis) + number of models + 1 (synthesis)\\n"
+            "- Models can have stances (for/against/neutral) for structured debate\\n"
+            "- Same model can be used multiple times with different stances\\n"
+            "- Each model + stance combination must be unique\\n\\n"
+            "Perfect for: complex decisions, architectural choices, feature proposals, "
+            "technology evaluations, strategic planning."
         )
 
-    def get_input_schema(self) -> dict[str, Any]:
-        schema = {
-            "type": "object",
-            "properties": {
-                "prompt": {
-                    "type": "string",
-                    "description": CONSENSUS_FIELD_DESCRIPTIONS["prompt"],
-                },
-                "models": {
-                    "type": "array",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "model": {
-                                "type": "string",
-                                "description": CONSENSUS_FIELD_DESCRIPTIONS["model_config_model"],
-                            },
-                            "stance": {
-                                "type": "string",
-                                "enum": ["for", "support", "favor", "against", "oppose", "critical", "neutral"],
-                                "description": CONSENSUS_FIELD_DESCRIPTIONS["model_config_stance_schema"],
-                                "default": "neutral",
-                            },
-                            "stance_prompt": {
-                                "type": "string",
-                                "description": CONSENSUS_FIELD_DESCRIPTIONS["model_config_stance_prompt"],
-                            },
-                        },
-                        "required": ["model"],
-                    },
-                    "description": CONSENSUS_FIELD_DESCRIPTIONS["models"],
-                },
-                "files": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": CONSENSUS_FIELD_DESCRIPTIONS["files"],
-                },
-                "images": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": CONSENSUS_FIELD_DESCRIPTIONS["images"],
-                },
-                "focus_areas": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": CONSENSUS_FIELD_DESCRIPTIONS["focus_areas"],
-                },
-                "temperature": {
-                    "type": "number",
-                    "description": "Temperature (0-1, default 0.2 for consistency)",
-                    "minimum": 0,
-                    "maximum": 1,
-                    "default": self.get_default_temperature(),
-                },
-                "thinking_mode": {
-                    "type": "string",
-                    "enum": ["minimal", "low", "medium", "high", "max"],
-                    "description": (
-                        "Thinking depth: minimal (0.5% of model max), low (8%), medium (33%), "
-                        "high (67%), max (100% of model max)"
-                    ),
-                },
-                "use_websearch": {
-                    "type": "boolean",
-                    "description": (
-                        "Enable web search for documentation, best practices, and current information. "
-                        "Particularly useful for: brainstorming sessions, architectural design discussions, "
-                        "exploring industry best practices, working with specific frameworks/technologies, "
-                        "researching solutions to complex problems, or when current documentation and "
-                        "community insights would enhance the analysis."
-                    ),
-                    "default": True,
-                },
-                "continuation_id": {
-                    "type": "string",
-                    "description": (
-                        "Thread continuation ID for multi-turn conversations. Can be used to continue "
-                        "conversations across different tools. Only provide this if continuing a previous "
-                        "conversation thread."
-                    ),
-                },
-            },
-            "required": ["prompt", "models"],
-        }
-
-        return schema
-
     def get_system_prompt(self) -> str:
-        return CONSENSUS_PROMPT
+        # For Claude's initial analysis, use a neutral version of the consensus prompt
+        return CONSENSUS_PROMPT.replace(
+            "{stance_prompt}",
+            """BALANCED PERSPECTIVE
+
+Provide objective analysis considering both positive and negative aspects. However, if there is overwhelming evidence
+that the proposal clearly leans toward being exceptionally good or particularly problematic, you MUST accurately
+reflect this reality. Being "balanced" means being truthful about the weight of evidence, not artificially creating
+50/50 splits when the reality is 90/10.
+
+Your analysis should:
+- Present all significant pros and cons discovered
+- Weight them according to actual impact and likelihood
+- If evidence strongly favors one conclusion, clearly state this
+- Provide proportional coverage based on the strength of arguments
+- Help the questioner see the true balance of considerations
+
+Remember: Artificial balance that misrepresents reality is not helpful. True balance means accurate representation
+of the evidence, even when it strongly points in one direction.""",
+        )
 
     def get_default_temperature(self) -> float:
-        return 0.2  # Lower temperature for more consistent consensus responses
+        return TEMPERATURE_ANALYTICAL
 
-    def get_model_category(self) -> "ToolModelCategory":
-        """Consensus uses extended reasoning models for deep analysis"""
+    def get_model_category(self) -> ToolModelCategory:
+        """Consensus workflow requires extended reasoning"""
         from tools.models import ToolModelCategory
 
         return ToolModelCategory.EXTENDED_REASONING
 
-    def get_request_model(self):
+    def get_workflow_request_model(self):
+        """Return the consensus workflow-specific request model."""
         return ConsensusRequest
 
-    def format_conversation_turn(self, turn) -> list[str]:
+    def get_input_schema(self) -> dict[str, Any]:
+        """Generate input schema for consensus workflow."""
+        from .workflow.schema_builders import WorkflowSchemaBuilder
+
+        # Consensus workflow-specific field overrides
+        consensus_field_overrides = {
+            "step": {
+                "type": "string",
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["step"],
+            },
+            "step_number": {
+                "type": "integer",
+                "minimum": 1,
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["step_number"],
+            },
+            "total_steps": {
+                "type": "integer",
+                "minimum": 1,
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["total_steps"],
+            },
+            "next_step_required": {
+                "type": "boolean",
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["next_step_required"],
+            },
+            "findings": {
+                "type": "string",
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["findings"],
+            },
+            "relevant_files": {
+                "type": "array",
+                "items": {"type": "string"},
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["relevant_files"],
+            },
+            "models": {
+                "type": "array",
+                "items": {
+                    "type": "object",
+                    "properties": {
+                        "model": {"type": "string"},
+                        "stance": {"type": "string", "enum": ["for", "against", "neutral"], "default": "neutral"},
+                        "stance_prompt": {"type": "string"},
+                    },
+                    "required": ["model"],
+                },
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["models"],
+            },
+            "current_model_index": {
+                "type": "integer",
+                "minimum": 0,
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["current_model_index"],
+            },
+            "model_responses": {
+                "type": "array",
+                "items": {"type": "object"},
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["model_responses"],
+            },
+            "images": {
+                "type": "array",
+                "items": {"type": "string"},
+                "description": CONSENSUS_WORKFLOW_FIELD_DESCRIPTIONS["images"],
+            },
+        }
+
+        # Build schema without standard workflow fields we don't use
+        schema = WorkflowSchemaBuilder.build_schema(
+            tool_specific_fields=consensus_field_overrides,
+            model_field_schema=self.get_model_field_schema(),
+            auto_mode=self.is_effective_auto_mode(),
+            tool_name=self.get_name(),
+        )
+
+        # Remove unused workflow fields
+        if "properties" in schema:
+            for field in [
+                "files_checked",
+                "relevant_context",
+                "issues_found",
+                "hypothesis",
+                "backtrack_from_step",
+                "confidence",  # Not used in consensus workflow
+                "temperature",  # Not used in consensus workflow
+                "thinking_mode",  # Not used in consensus workflow
+                "use_websearch",  # Not used in consensus workflow
+                "relevant_files",  # Not used in consensus workflow
+            ]:
+                schema["properties"].pop(field, None)
+
+        return schema
+
+    def get_required_actions(
+        self, step_number: int, confidence: str, findings: str, total_steps: int
+    ) -> list[str]:  # noqa: ARG002
+        """Define required actions for each consensus phase.
+
+        Note: confidence parameter is kept for compatibility with base class but not used.
         """
-        Format consensus turns with individual model responses for better readability.
-
-        This custom formatting shows the individual model responses that were
-        synthesized into the consensus, making it easier to understand the
-        reasoning behind the final recommendation.
-        """
-        parts = []
-
-        # Add files context if present
-        if turn.files:
-            parts.append(f"Files used in this turn: {', '.join(turn.files)}")
-            parts.append("")
-
-        # Check if this is a consensus turn with individual responses
-        if turn.model_metadata and turn.model_metadata.get("individual_responses"):
-            individual_responses = turn.model_metadata["individual_responses"]
-
-            # Add consensus header
-            models_consulted = []
-            for resp in individual_responses:
-                model = resp["model"]
-                stance = resp.get("stance", "neutral")
-                if stance != "neutral":
-                    models_consulted.append(f"{model}:{stance}")
-                else:
-                    models_consulted.append(model)
-
-            parts.append(f"Models consulted: {', '.join(models_consulted)}")
-            parts.append("")
-            parts.append("=== INDIVIDUAL MODEL RESPONSES ===")
-            parts.append("")
-
-            # Add each successful model response
-            for i, response in enumerate(individual_responses):
-                model_name = response["model"]
-                stance = response.get("stance", "neutral")
-                verdict = response["verdict"]
-
-                stance_label = f"({stance.title()} Stance)" if stance != "neutral" else "(Neutral Analysis)"
-                parts.append(f"**{model_name.upper()} {stance_label}**:")
-                parts.append(verdict)
-
-                if i < len(individual_responses) - 1:
-                    parts.append("")
-                    parts.append("---")
-                parts.append("")
-
-            parts.append("=== END INDIVIDUAL RESPONSES ===")
-            parts.append("")
-            parts.append("Claude's Synthesis:")
-
-        # Add the actual content
-        parts.append(turn.content)
-
-        return parts
-
-    def _normalize_stance(self, stance: Optional[str]) -> str:
-        """Normalize stance to canonical form."""
-        if not stance:
-            return "neutral"
-
-        stance = stance.lower()
-
-        # Define stance synonyms
-        supportive_stances = {"for", "support", "favor"}
-        critical_stances = {"against", "oppose", "critical"}
-
-        # Map synonyms to canonical stance
-        if stance in supportive_stances:
-            return "for"
-        elif stance in critical_stances:
-            return "against"
-        elif stance == "neutral":
-            return "neutral"
+        if step_number == 1:
+            # Claude's initial analysis
+            return [
+                "You've provided your initial analysis. The tool will now consult other models.",
+                "Wait for the next step to receive the first model's response.",
+            ]
+        elif step_number < total_steps - 1:
+            # Processing individual model responses
+            return [
+                "Review the model response provided in this step",
+                "Note key agreements and disagreements with previous analyses",
+                "Wait for the next model's response",
+            ]
         else:
-            # Unknown stances default to neutral for robustness
-            logger.warning(
-                f"Unknown stance '{stance}' provided, defaulting to 'neutral'. Valid stances: {', '.join(sorted(supportive_stances | critical_stances))}, or 'neutral'"
+            # Ready for final synthesis
+            return [
+                "All models have been consulted",
+                "Synthesize all perspectives into a comprehensive recommendation",
+                "Identify key points of agreement and disagreement",
+                "Provide clear, actionable guidance based on the consensus",
+            ]
+
+    def should_call_expert_analysis(self, consolidated_findings, request=None) -> bool:
+        """Consensus workflow doesn't use traditional expert analysis - it consults models step by step."""
+        return False
+
+    def prepare_expert_analysis_context(self, consolidated_findings) -> str:
+        """Not used in consensus workflow."""
+        return ""
+
+    def requires_expert_analysis(self) -> bool:
+        """Consensus workflow handles its own model consultations."""
+        return False
+
+    # Hook method overrides for consensus-specific behavior
+
+    def prepare_step_data(self, request) -> dict:
+        """Prepare consensus-specific step data."""
+        step_data = {
+            "step": request.step,
+            "step_number": request.step_number,
+            "findings": request.findings,
+            "files_checked": [],  # Not used
+            "relevant_files": request.relevant_files or [],
+            "relevant_context": [],  # Not used
+            "issues_found": [],  # Not used
+            "confidence": "exploring",  # Not used, kept for compatibility
+            "hypothesis": None,  # Not used
+            "images": request.images or [],  # Now used for visual context
+        }
+        return step_data
+
+    async def handle_work_completion(self, response_data: dict, request, arguments: dict) -> dict:  # noqa: ARG002
+        """Handle consensus workflow completion - no expert analysis, just final synthesis."""
+        response_data["consensus_complete"] = True
+        response_data["status"] = "consensus_workflow_complete"
+
+        # Prepare final synthesis data
+        response_data["complete_consensus"] = {
+            "initial_prompt": self.initial_prompt,
+            "models_consulted": [m["model"] + ":" + m.get("stance", "neutral") for m in self.accumulated_responses],
+            "total_responses": len(self.accumulated_responses),
+            "consensus_confidence": "high",  # Consensus complete
+        }
+
+        response_data["next_steps"] = (
+            "CONSENSUS GATHERING IS COMPLETE. You MUST now synthesize all perspectives and present:\n"
+            "1. Key points of AGREEMENT across models\n"
+            "2. Key points of DISAGREEMENT and why they differ\n"
+            "3. Your final consolidated recommendation\n"
+            "4. Specific, actionable next steps for implementation\n"
+            "5. Critical risks or concerns that must be addressed"
+        )
+
+        return response_data
+
+    def handle_work_continuation(self, response_data: dict, request) -> dict:
+        """Handle continuation between consensus steps."""
+        current_idx = request.current_model_index or 0
+
+        if request.step_number == 1:
+            # After Claude's initial analysis, prepare to consult first model
+            response_data["status"] = "consulting_models"
+            response_data["next_model"] = self.models_to_consult[0] if self.models_to_consult else None
+            response_data["next_steps"] = (
+                "Your initial analysis is complete. The tool will now consult the specified models."
             )
-            return "neutral"
+        elif current_idx < len(self.models_to_consult):
+            next_model = self.models_to_consult[current_idx]
+            response_data["status"] = "consulting_next_model"
+            response_data["next_model"] = next_model
+            response_data["models_remaining"] = len(self.models_to_consult) - current_idx
+            response_data["next_steps"] = f"Model consultation in progress. Next: {next_model['model']}"
+        else:
+            response_data["status"] = "ready_for_synthesis"
+            response_data["next_steps"] = "All models consulted. Ready for final synthesis."
 
-    def _validate_model_combinations(self, model_configs: list[ModelConfig]) -> tuple[list[ModelConfig], list[str]]:
-        """Validate model configurations and enforce limits.
+        return response_data
 
-        Returns:
-            tuple: (valid_configs, skipped_entries)
-            - Each model+stance combination can appear max 2 times
-            - Same model+stance limited to 2 instances
-        """
-        valid_configs = []
-        skipped_entries = []
-        combination_counts = {}  # Track (model, stance) -> count
+    async def execute_workflow(self, arguments: dict[str, Any]) -> list:
+        """Override execute_workflow to handle model consultations between steps."""
 
-        for config in model_configs:
-            try:
-                # Normalize stance
-                normalized_stance = self._normalize_stance(config.stance)
+        # Store arguments
+        self._current_arguments = arguments
 
-                # Create normalized config
-                normalized_config = ModelConfig(
-                    model=config.model, stance=normalized_stance, stance_prompt=config.stance_prompt
-                )
+        # Validate request
+        request = self.get_workflow_request_model()(**arguments)
 
-                combination_key = (config.model, normalized_stance)
-                current_count = combination_counts.get(combination_key, 0)
+        # On first step, store the models to consult
+        if request.step_number == 1:
+            self.initial_prompt = request.step
+            self.models_to_consult = request.models or []
+            self.accumulated_responses = []
+            # Set total steps: 1 (Claude) + len(models) + 1 (synthesis)
+            request.total_steps = 1 + len(self.models_to_consult) + 1
 
-                if current_count >= DEFAULT_CONSENSUS_MAX_INSTANCES_PER_COMBINATION:
-                    # Already have max instances of this model+stance combination
-                    skipped_entries.append(
-                        f"{config.model}:{normalized_stance} (max {DEFAULT_CONSENSUS_MAX_INSTANCES_PER_COMBINATION} instances)"
+        # If this is a model consultation step (2 through total_steps-1)
+        elif request.step_number > 1 and request.step_number < request.total_steps:
+            # Get the current model to consult
+            model_idx = request.current_model_index or 0
+            if model_idx < len(self.models_to_consult):
+                # Consult the model
+                model_response = await self._consult_model(self.models_to_consult[model_idx], request)
+
+                # Add to accumulated responses
+                self.accumulated_responses.append(model_response)
+
+                # Include the model response in the step data
+                response_data = {
+                    "status": "model_consulted",
+                    "step_number": request.step_number,
+                    "total_steps": request.total_steps,
+                    "model_consulted": model_response["model"],
+                    "model_stance": model_response.get("stance", "neutral"),
+                    "model_response": model_response,
+                    "current_model_index": model_idx + 1,
+                    "next_step_required": request.step_number < request.total_steps - 1,
+                }
+
+                if request.step_number < request.total_steps - 1:
+                    response_data["next_steps"] = (
+                        f"Model {model_response['model']} has provided its {model_response.get('stance', 'neutral')} "
+                        f"perspective. Please analyze this response and call {self.get_name()} again with:\n"
+                        f"- step_number: {request.step_number + 1}\n"
+                        f"- findings: Summarize key points from this model's response\n"
+                        f"- current_model_index: {model_idx + 1}\n"
+                        f"- model_responses: (append this response to the list)"
                     )
-                    continue
+                else:
+                    response_data["next_steps"] = (
+                        "All models have been consulted. For the final step, synthesize all perspectives."
+                    )
 
-                combination_counts[combination_key] = current_count + 1
-                valid_configs.append(normalized_config)
+                return [TextContent(type="text", text=json.dumps(response_data, indent=2))]
 
-            except ValueError as e:
-                # Invalid stance or model
-                skipped_entries.append(f"{config.model} ({str(e)})")
-                continue
+        # Otherwise, use standard workflow execution
+        return await super().execute_workflow(arguments)
 
-        return valid_configs, skipped_entries
+    async def _consult_model(self, model_config: dict, request) -> dict:
+        """Consult a single model and return its response."""
+        try:
+            # Get the provider for this model
+            model_name = model_config["model"]
+            provider = self.get_model_provider(model_name)
 
-    def _get_stance_enhanced_prompt(self, stance: str, custom_stance_prompt: Optional[str] = None) -> str:
-        """Get the system prompt with stance injection based on the stance."""
-        base_prompt = self.get_system_prompt()
-
-        # If custom stance prompt is provided, use it instead of default
-        if custom_stance_prompt:
-            # Validate stance placeholder exists exactly once
-            if base_prompt.count("{stance_prompt}") != 1:
-                raise ValueError(
-                    "System prompt must contain exactly one '{stance_prompt}' placeholder, "
-                    f"found {base_prompt.count('{stance_prompt}')}"
+            # Prepare the prompt with any relevant files
+            prompt = self.initial_prompt
+            if request.relevant_files:
+                file_content, _ = self._prepare_file_content_for_prompt(
+                    request.relevant_files,
+                    request.continuation_id,
+                    "Context files",
                 )
+                if file_content:
+                    prompt = f"{prompt}\n\n=== CONTEXT FILES ===\n{file_content}\n=== END CONTEXT ==="
+
+            # Get stance-specific system prompt
+            stance = model_config.get("stance", "neutral")
+            stance_prompt = model_config.get("stance_prompt")
+            system_prompt = self._get_stance_enhanced_prompt(stance, stance_prompt)
+
+            # Call the model
+            response = provider.generate_content(
+                prompt=prompt,
+                model_name=model_name,
+                system_prompt=system_prompt,
+                temperature=0.2,  # Low temperature for consistency
+                thinking_mode="medium",
+                images=request.images if request.images else None,
+            )
+
+            return {
+                "model": model_name,
+                "stance": stance,
+                "status": "success",
+                "verdict": response.content,
+                "metadata": {
+                    "provider": provider.get_provider_type().value,
+                },
+            }
+
+        except Exception as e:
+            logger.exception("Error consulting model %s", model_config)
+            return {
+                "model": model_config.get("model", "unknown"),
+                "stance": model_config.get("stance", "neutral"),
+                "status": "error",
+                "error": str(e),
+            }
+
+    def _get_stance_enhanced_prompt(self, stance: str, custom_stance_prompt: str | None = None) -> str:
+        """Get the system prompt with stance injection."""
+        base_prompt = CONSENSUS_PROMPT
+
+        if custom_stance_prompt:
             return base_prompt.replace("{stance_prompt}", custom_stance_prompt)
 
         stance_prompts = {
@@ -379,7 +561,9 @@ YOUR SUPPORTIVE ANALYSIS SHOULD:
 - Suggest optimizations that enhance value
 - Present realistic implementation pathways
 
-Remember: Being "for" means finding the BEST possible version of the idea IF it has merit, not blindly supporting bad ideas.""",
+Remember: Being "for" means finding the BEST possible version of the idea IF it has merit, not blindly supporting bad """
+            "ideas."
+            "",
             "against": """CRITICAL PERSPECTIVE WITH RESPONSIBILITY
 
 You are tasked with critiquing this proposal, but with ESSENTIAL BOUNDARIES:
@@ -403,7 +587,9 @@ YOUR CRITICAL ANALYSIS SHOULD:
 - Highlight potential negative consequences
 - Question assumptions that may be flawed
 
-Remember: Being "against" means rigorous scrutiny to ensure quality, not undermining good ideas that deserve support.""",
+Remember: Being "against" means rigorous scrutiny to ensure quality, not undermining good ideas that deserve """
+            "support."
+            "",
             "neutral": """BALANCED PERSPECTIVE
 
 Provide objective analysis considering both positive and negative aspects. However, if there is overwhelming evidence
@@ -423,371 +609,33 @@ of the evidence, even when it strongly points in one direction.""",
         }
 
         stance_prompt = stance_prompts.get(stance, stance_prompts["neutral"])
-
-        # Validate stance placeholder exists exactly once
-        if base_prompt.count("{stance_prompt}") != 1:
-            raise ValueError(
-                "System prompt must contain exactly one '{stance_prompt}' placeholder, "
-                f"found {base_prompt.count('{stance_prompt}')}"
-            )
-
-        # Inject stance into the system prompt
         return base_prompt.replace("{stance_prompt}", stance_prompt)
 
-    def _get_single_response(
-        self, provider, model_config: ModelConfig, prompt: str, request: ConsensusRequest
-    ) -> dict[str, Any]:
-        """Get response from a single model - synchronous method."""
-        logger.debug(f"Getting response from {model_config.model} with stance '{model_config.stance}'")
+    def customize_workflow_response(self, response_data: dict, request) -> dict:
+        """Customize response for consensus workflow."""
+        # Store model responses in the response for tracking
+        if self.accumulated_responses:
+            response_data["accumulated_responses"] = self.accumulated_responses
 
-        try:
-            # Provider.generate_content is synchronous, not async
-            response = provider.generate_content(
-                prompt=prompt,
-                model_name=model_config.model,
-                system_prompt=self._get_stance_enhanced_prompt(model_config.stance, model_config.stance_prompt),
-                temperature=getattr(request, "temperature", None) or self.get_default_temperature(),
-                thinking_mode=getattr(request, "thinking_mode", "medium"),
-                images=getattr(request, "images", None) or [],
-            )
-            return {
-                "model": model_config.model,
-                "stance": model_config.stance,
-                "status": "success",
-                "verdict": response.content,  # Contains structured Markdown
-                "metadata": {
-                    "provider": getattr(provider.get_provider_type(), "value", provider.get_provider_type()),
-                    "usage": response.usage if hasattr(response, "usage") else None,
-                    "custom_stance_prompt": bool(model_config.stance_prompt),
-                },
-            }
-        except Exception as e:
-            logger.error(f"Error getting response from {model_config.model}:{model_config.stance}: {str(e)}")
-            return {"model": model_config.model, "stance": model_config.stance, "status": "error", "error": str(e)}
+        # Add consensus-specific fields
+        if request.step_number == 1:
+            response_data["consensus_workflow_status"] = "initial_analysis_complete"
+        elif request.step_number < request.total_steps - 1:
+            response_data["consensus_workflow_status"] = "consulting_models"
+        else:
+            response_data["consensus_workflow_status"] = "ready_for_synthesis"
 
-    def _get_consensus_responses(
-        self, provider_configs: list[tuple], prompt: str, request: ConsensusRequest
-    ) -> list[dict[str, Any]]:
-        """Execute all model requests sequentially - purely synchronous like other tools."""
+        return response_data
 
-        logger.debug(f"Processing {len(provider_configs)} models sequentially")
-        responses = []
+    def store_initial_issue(self, step_description: str):
+        """Store initial prompt for model consultations."""
+        self.initial_prompt = step_description
 
-        for i, (provider, model_config) in enumerate(provider_configs):
-            try:
-                logger.debug(
-                    f"Processing {model_config.model}:{model_config.stance} sequentially ({i+1}/{len(provider_configs)})"
-                )
+    # Required abstract methods from BaseTool
+    def get_request_model(self):
+        """Return the consensus workflow-specific request model."""
+        return ConsensusRequest
 
-                # Direct synchronous call - matches pattern of other tools
-                response = self._get_single_response(provider, model_config, prompt, request)
-                responses.append(response)
-
-            except Exception as e:
-                logger.error(f"Failed to get response from {model_config.model}:{model_config.stance}: {str(e)}")
-                responses.append(
-                    {
-                        "model": model_config.model,
-                        "stance": model_config.stance,
-                        "status": "error",
-                        "error": f"Unhandled exception: {str(e)}",
-                    }
-                )
-
-        logger.debug(f"Sequential processing completed for {len(responses)} models")
-        return responses
-
-    def _format_consensus_output(self, responses: list[dict[str, Any]], skipped_entries: list[str]) -> str:
-        """Format the consensus responses into structured output for Claude."""
-
-        logger.debug(f"Formatting consensus output for {len(responses)} responses")
-
-        # Separate successful and failed responses
-        successful_responses = [r for r in responses if r["status"] == "success"]
-        failed_responses = [r for r in responses if r["status"] == "error"]
-
-        logger.debug(f"Successful responses: {len(successful_responses)}, Failed: {len(failed_responses)}")
-
-        # Prepare the structured output (minimize size for MCP stability)
-        models_used = [
-            f"{r['model']}:{r['stance']}" if r["stance"] != "neutral" else r["model"] for r in successful_responses
-        ]
-        models_errored = [
-            f"{r['model']}:{r['stance']}" if r["stance"] != "neutral" else r["model"] for r in failed_responses
-        ]
-
-        # Prepare clean responses without truncation
-        clean_responses = []
-        for r in responses:
-            if r["status"] == "success":
-                clean_responses.append(
-                    {
-                        "model": r["model"],
-                        "stance": r["stance"],
-                        "status": r["status"],
-                        "verdict": r.get("verdict", ""),
-                        "metadata": r.get("metadata", {}),
-                    }
-                )
-            else:
-                clean_responses.append(
-                    {
-                        "model": r["model"],
-                        "stance": r["stance"],
-                        "status": r["status"],
-                        "error": r.get("error", "Unknown error"),
-                    }
-                )
-
-        output_data = {
-            "status": "consensus_success" if successful_responses else "consensus_failed",
-            "models_used": models_used,
-            "models_skipped": skipped_entries,
-            "models_errored": models_errored,
-            "responses": clean_responses,
-            "next_steps": self._get_synthesis_guidance(successful_responses, failed_responses),
-        }
-
-        return json.dumps(output_data, indent=2)
-
-    def _get_synthesis_guidance(
-        self, successful_responses: list[dict[str, Any]], failed_responses: list[dict[str, Any]]
-    ) -> str:
-        """Generate guidance for Claude on how to synthesize the consensus results."""
-
-        if not successful_responses:
-            return (
-                "No models provided successful responses. Please retry with different models or "
-                "check the error messages for guidance on resolving the issues."
-            )
-
-        if len(successful_responses) == 1:
-            return (
-                "Only one model provided a successful response. Synthesize based on the available "
-                "perspective and indicate areas where additional expert input would be valuable "
-                "due to the limited consensus data."
-            )
-
-        # Multiple successful responses - provide comprehensive synthesis guidance
-        stance_counts = {"for": 0, "against": 0, "neutral": 0}
-        for resp in successful_responses:
-            stance = resp.get("stance", "neutral")
-            stance_counts[stance] = stance_counts.get(stance, 0) + 1
-
-        guidance = (
-            "Claude, synthesize these perspectives by first identifying the key points of "
-            "**agreement** and **disagreement** between the models. Then provide your final, "
-            "consolidated recommendation, explaining how you weighed the different opinions and "
-            "why your proposed solution is the most balanced approach. Explicitly address the "
-            "most critical risks raised by each model and provide actionable next steps for implementation."
-        )
-
-        if failed_responses:
-            guidance += (
-                f" Note: {len(failed_responses)} model(s) failed to respond - consider this "
-                "partial consensus and indicate where additional expert input would strengthen the analysis."
-            )
-
-        return guidance
-
-    async def prepare_prompt(self, request: ConsensusRequest) -> str:
-        """Prepare the consensus prompt with context files and focus areas."""
-        # Check for prompt.txt in files
-        prompt_content, updated_files = self.handle_prompt_file(request.files)
-
-        # Use prompt.txt content if available, otherwise use the prompt field
-        user_content = prompt_content if prompt_content else request.prompt
-
-        # Check user input size at MCP transport boundary (before adding internal content)
-        size_check = self.check_prompt_size(user_content)
-        if size_check:
-            # Need to return error, but prepare_prompt returns str
-            # Use exception to handle this cleanly
-            from tools.models import ToolOutput
-
-            raise ValueError(f"MCP_SIZE_CHECK:{ToolOutput(**size_check).model_dump_json()}")
-
-        # Update request files list
-        if updated_files is not None:
-            request.files = updated_files
-
-        # Add focus areas if specified
-        if request.focus_areas:
-            focus_areas_text = "\n\nSpecific focus areas for this analysis:\n" + "\n".join(
-                f"- {area}" for area in request.focus_areas
-            )
-            user_content += focus_areas_text
-
-        # Add context files if provided (using centralized file handling with filtering)
-        if request.files:
-            file_content, processed_files = self._prepare_file_content_for_prompt(
-                request.files, request.continuation_id, "Context files"
-            )
-            self._actually_processed_files = processed_files
-            if file_content:
-                user_content = f"{user_content}\n\n=== CONTEXT FILES ===\n{file_content}\n=== END CONTEXT ===="
-
-        # Check token limits
-        self._validate_token_limit(user_content, "Content")
-
-        return user_content
-
-    async def execute(self, arguments: dict[str, Any]) -> list[TextContent]:
-        """Execute consensus gathering from multiple models."""
-
-        # Store arguments for base class methods
-        self._current_arguments = arguments
-
-        # Validate and create request
-        request = ConsensusRequest(**arguments)
-
-        # Validate model configurations and enforce limits
-        valid_configs, skipped_entries = self._validate_model_combinations(request.models)
-
-        if not valid_configs:
-            error_output = {
-                "status": "consensus_failed",
-                "error": "No valid model configurations after validation",
-                "models_skipped": skipped_entries,
-                "next_steps": "Please provide valid model configurations with proper model names and stance values.",
-            }
-            return [TextContent(type="text", text=json.dumps(error_output, indent=2))]
-
-        # Set up a dummy model context for consensus since we handle multiple models
-        # This is needed for base class methods like prepare_prompt to work
-        if not hasattr(self, "_model_context") or not self._model_context:
-            from utils.model_context import ModelContext
-
-            # Use the first model as the representative for token calculations
-            first_model = valid_configs[0].model if valid_configs else "flash"
-            self._model_context = ModelContext(first_model)
-
-        # Handle conversation continuation if specified
-        if request.continuation_id:
-            from utils.conversation_memory import build_conversation_history, get_thread
-
-            thread_context = get_thread(request.continuation_id)
-            if thread_context:
-                # Build conversation history using the same pattern as other tools
-                conversation_context, _ = build_conversation_history(thread_context, self._model_context)
-                if conversation_context:
-                    # Add conversation context to the beginning of the prompt
-                    enhanced_prompt = f"{conversation_context}\n\n{request.prompt}"
-                    request.prompt = enhanced_prompt
-
-        # Prepare the consensus prompt
-        consensus_prompt = await self.prepare_prompt(request)
-
-        # Get providers for valid model configurations with caching to avoid duplicate lookups
-        provider_configs = []
-        provider_cache = {}  # Cache to avoid duplicate provider lookups
-
-        for model_config in valid_configs:
-            try:
-                # Check cache first
-                if model_config.model in provider_cache:
-                    provider = provider_cache[model_config.model]
-                else:
-                    # Look up provider and cache it
-                    provider = self.get_model_provider(model_config.model)
-                    provider_cache[model_config.model] = provider
-
-                provider_configs.append((provider, model_config))
-            except Exception as e:
-                # Track failed models
-                model_display = (
-                    f"{model_config.model}:{model_config.stance}"
-                    if model_config.stance != "neutral"
-                    else model_config.model
-                )
-                skipped_entries.append(f"{model_display} (provider not available: {str(e)})")
-
-        if not provider_configs:
-            error_output = {
-                "status": "consensus_failed",
-                "error": "No model providers available",
-                "models_skipped": skipped_entries,
-                "next_steps": "Please check that the specified models have configured API keys and are available.",
-            }
-            return [TextContent(type="text", text=json.dumps(error_output, indent=2))]
-
-        # Send to all models sequentially (purely synchronous like other tools)
-        logger.debug(f"Sending consensus request to {len(provider_configs)} models")
-        responses = self._get_consensus_responses(provider_configs, consensus_prompt, request)
-        logger.debug(f"Received {len(responses)} responses from consensus models")
-
-        # Enforce minimum success requirement - must have at least 1 successful response
-        successful_responses = [r for r in responses if r["status"] == "success"]
-        if not successful_responses:
-            error_output = {
-                "status": "consensus_failed",
-                "error": "All model calls failed - no successful responses received",
-                "models_skipped": skipped_entries,
-                "models_errored": [
-                    f"{r['model']}:{r['stance']}" if r["stance"] != "neutral" else r["model"]
-                    for r in responses
-                    if r["status"] == "error"
-                ],
-                "next_steps": "Please retry with different models or check the error messages for guidance on resolving the issues.",
-            }
-            return [TextContent(type="text", text=json.dumps(error_output, indent=2))]
-
-        logger.debug("About to format consensus output for MCP response")
-
-        # Structure the output and store in conversation memory
-        consensus_output = self._format_consensus_output(responses, skipped_entries)
-
-        # Log response size for debugging
-        output_size = len(consensus_output)
-        logger.debug(f"Consensus output size: {output_size:,} characters")
-
-        # Store in conversation memory if continuation_id is provided
-        if request.continuation_id:
-            self.store_conversation_turn(
-                request.continuation_id,
-                consensus_output,
-                request.files,
-                request.images,
-                responses,  # Store individual responses in metadata
-                skipped_entries,
-            )
-
-        return [TextContent(type="text", text=consensus_output)]
-
-    def store_conversation_turn(
-        self,
-        continuation_id: str,
-        output: str,
-        files: list[str],
-        images: list[str],
-        responses: list[dict[str, Any]],
-        skipped_entries: list[str],
-    ):
-        """Store consensus turn in conversation memory with special metadata."""
-        from utils.conversation_memory import add_turn
-
-        # Filter successful and failed responses
-        successful_responses = [r for r in responses if r["status"] == "success"]
-        failed_responses = [r for r in responses if r["status"] == "error"]
-
-        # Prepare metadata for conversation storage
-        metadata = {
-            "tool_type": "consensus",
-            "models_used": [r["model"] for r in successful_responses],
-            "models_skipped": skipped_entries,
-            "models_errored": [r["model"] for r in failed_responses],
-            "individual_responses": successful_responses,  # Only store successful responses
-        }
-
-        # Store the turn with special consensus metadata - add_turn is synchronous
-        add_turn(
-            thread_id=continuation_id,
-            role="assistant",
-            content=output,
-            files=files or [],
-            images=images or [],
-            tool_name="consensus",
-            model_provider="consensus",  # Special provider name
-            model_name="consensus",  # Special model name
-            model_metadata=metadata,
-        )
+    async def prepare_prompt(self, request) -> str:  # noqa: ARG002
+        """Not used - workflow tools use execute_workflow()."""
+        return ""  # Workflow tools use execute_workflow() directly
